+++ docs/source/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 * Updated Modernizr to v2.8.3
 * Updated jquery.dataTables to v1.10.1
 * Added jquery.dataTables minified CSS.
+* Added djfrontend_jquery_datatables_themeroller template tag.
 
 1.3.0
 ------